+++ packages/roosterjs-editor-dom/lib/contentTraverser/ContentTraverser.ts
@@ -1,7 +1,8 @@
 import BodyScoper from './BodyScoper';
 import EmptyInlineElement from '../inlineElements/EmptyInlineElement';
 import getBlockElementAtNode from '../blockElements/getBlockElementAtNode';
-import getNextPreviousInlineElement from '../inlineElements/getNextPreviousInlineElement';
+import getInlineElementAtNode from '../inlineElements/getInlineElementAtNode';
+import PartialInlineElement from '../inlineElements/PartialInlineElement';
 import SelectionBlockScoper from './SelectionBlockScoper';
 import SelectionScoper from './SelectionScoper';
 import TraversingScoper from './TraversingScoper';
@@ -170,4 +171,27 @@ export default class ContentTraverser {
 
         return null;
     }
+}
+
+function getNextPreviousInlineElement(
+    rootNode: Node,
+    current: InlineElement,
+    isNext: boolean
+): InlineElement {
+    if (!current) {
+        return null;
+    }
+    if (current instanceof PartialInlineElement) {
+        // if current is partial, get the the othe half of the inline unless it is no more
+        let result = isNext ? current.nextInlineElement : current.previousInlineElement;
+
+        if (result) {
+            return result;
+        }
+    }
+
+    // Get a leaf node after startNode and use that base to find next inline
+    let startNode = current.getContainerNode();
+    startNode = getLeafSibling(rootNode, startNode, isNext);
+    return getInlineElementAtNode(rootNode, startNode);
 }
